Ejercicio interseccion de circulos
