+++ Fig_Geometricas/fig_geometricas_her/CMakeLists.txt
@@ -1,4 +1,4 @@
 cmake_minimum_required(VERSION 3.10)
 project(fig_geometricas_her)
-add_executable(${PROJECT_NAME} main.cpp Geometrica.h Geometrica.cpp Circulo.h Circulo.cpp Cuadrado.h
+add_executable(${PROJECT_NAME} main-tarea.cpp Geometrica.h Geometrica.cpp Circulo-tarea.h Circulo-tarea.cpp Cuadrado.h
     Cuadrado.cpp Triangulo.h Triangulo.cpp Pentagono.h Pentagono.cpp )
